requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-environ==0.12.0
idna==3.10
pillow==11.2.1
python-http-client==3.3.7
pytz==2025.2
requests==2.32.3
sendgrid==6.11.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
stripe==12.1.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
